Fikser endring av harAktivitetsplikt hvis det eksisterer aktiviteter
