+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/aktivitetsplikt/AktivitetspliktVurdering.tsx
@@ -1,12 +1,13 @@
-import { BodyShort, Box, Button, Heading, HStack, Label, Radio, Textarea, VStack } from '@navikt/ds-react'
-import React, { useEffect, useState } from 'react'
+import { Box, Heading, VStack } from '@navikt/ds-react'
+import React, { useEffect } from 'react'
 import { useApiCall } from '~shared/hooks/useApiCall'
 import {
   AktivitetspliktOppgaveVurderingType,
   AktivitetspliktUnntakType,
+  IAktivitetspliktUnntak,
   IAktivitetspliktVurderingNyDto,
 } from '~shared/types/Aktivitetsplikt'
-import { hentAktivitspliktVurderingForBehandling, redigerUnntakForBehandling } from '~shared/api/aktivitetsplikt'
+import { hentAktivitspliktVurderingForBehandling } from '~shared/api/aktivitetsplikt'
 import { IDetaljertBehandling } from '~shared/types/IDetaljertBehandling'
 import { useInnloggetSaksbehandler } from '~components/behandling/useInnloggetSaksbehandler'
 import { behandlingErRedigerbar } from '~components/behandling/felles/utils'
@@ -20,203 +21,16 @@ import {
 } from '~store/reducers/AktivitetspliktBehandlingReducer'
 import { VurderAktivitetspliktWrapperBehandling } from '~components/behandling/aktivitetsplikt/VurderAktivitetspliktWrapperBehandling'
 import { isBefore, subMonths } from 'date-fns'
-import { JaNei } from '~shared/types/ISvar'
 import { HjemmelLenke } from '~components/behandling/felles/HjemmelLenke'
-import { useAppDispatch } from '~store/Store'
-import { useForm } from 'react-hook-form'
-import { ControlledRadioGruppe } from '~shared/components/radioGruppe/ControlledRadioGruppe'
+import { HarBrukerVarigUnntak } from '~components/behandling/aktivitetsplikt/unntak/HarBrukerVarigUnntak'
 
-const vurderingHarInnhold = (vurdering: IAktivitetspliktVurderingNyDto): boolean => {
+export const vurderingHarInnhold = (vurdering: IAktivitetspliktVurderingNyDto): boolean => {
   return !!vurdering.unntak.length || !!vurdering.aktivitet.length
 }
 
-const harVarigUnntak = (vurdering: IAktivitetspliktVurderingNyDto): boolean => {
-  return (
-    !!vurdering.unntak.length &&
-    !!vurdering.unntak.find(
-      (unntak) => unntak.unntak === AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT
-    )
-  )
-}
-
-const formvaluesFraVurdering = (vurdering: IAktivitetspliktVurderingNyDto): Partial<HarBrukerVarigUnntakFormdata> => ({
-  harAktivitetsplikt: harAktivitetspliktFraVurdering(vurdering),
-  beskrivelse: vurdering.unntak.find(
-    (u) => u.unntak === AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT
-  )?.beskrivelse,
-})
-
-function harAktivitetspliktFraVurdering(vurdering: IAktivitetspliktVurderingNyDto): JaNei | undefined {
-  if (!vurderingHarInnhold(vurdering)) {
-    return undefined
-  }
-  if (harVarigUnntak(vurdering)) {
-    return JaNei.NEI
-  } else {
-    return JaNei.JA
-  }
-}
-
-interface HarBrukerVarigUnntakFormdata {
-  harAktivitetsplikt: JaNei
-  beskrivelse: string
-}
-
-function HarBrukerVarigUnntak(props: { behandling: IDetaljertBehandling; doedsdato: Date; redigerbar: boolean }) {
-  const { behandling, redigerbar, doedsdato } = props
-  const dispatch = useAppDispatch()
-  const vurdering = useAktivitetspliktBehandlingState()
-  const [lagreUnntakVarigStatus, lagreUnntakVarig] = useApiCall(redigerUnntakForBehandling)
-  const [redigerer, setRedigerer] = useState<boolean>(!harAktivitetspliktFraVurdering(vurdering))
-
-  const {
-    handleSubmit,
-    register,
-    watch,
-    control,
-    reset,
-    formState: { errors },
-  } = useForm<HarBrukerVarigUnntakFormdata>({
-    defaultValues: formvaluesFraVurdering(vurdering),
-  })
-
-  useEffect(() => {
-    setRedigerer(!harAktivitetspliktFraVurdering(vurdering))
-    reset(formvaluesFraVurdering(vurdering))
-  }, [vurdering])
-
-  function avbrytRedigering() {
-    reset(formvaluesFraVurdering(vurdering))
-    setRedigerer(false)
-  }
-
-  const jaHarAktivitetsplikt = watch('harAktivitetsplikt') === JaNei.JA
-
-  const lagreVarigUnntak = (formdata: HarBrukerVarigUnntakFormdata) => {
-    if (formdata.harAktivitetsplikt === JaNei.NEI) {
-      lagreUnntakVarig(
-        {
-          sakId: behandling.sakId,
-          behandlingId: behandling.id,
-          request: {
-            id: undefined,
-            unntak: AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT,
-            fom: new Date().toISOString(),
-            tom: undefined,
-            beskrivelse: formdata.beskrivelse,
-          },
-        },
-        (data) => {
-          dispatch(setVurderingBehandling(data))
-          avbrytRedigering()
-        }
-      )
-    }
-  }
-
-  return redigerbar && redigerer ? (
-    <>
-      <form onSubmit={handleSubmit(lagreVarigUnntak)}>
-        <VStack maxWidth="32.5rem" gap="4">
-          <ControlledRadioGruppe
-            control={control}
-            name="harAktivitetsplikt"
-            legend="Har bruker aktivitetsplikt etter overgangsperioden?"
-            description={
-              <BodyShort>
-                Vurder om bruker fra 6 måneder etter dødsfallet må være i aktivitet, eller har varig unntak.
-              </BodyShort>
-            }
-            radios={
-              <>
-                <Radio size="small" value={JaNei.JA}>
-                  Ja, bruker har aktivitetsplikt
-                </Radio>
-                <Radio size="small" value={JaNei.NEI}>
-                  Nei, bruker er født i 1963 eller tidligere og har lav inntekt
-                </Radio>
-              </>
-            }
-            errorVedTomInput="Du må velge om bruker har aktivitetsplikt"
-          />
-          {watch('harAktivitetsplikt') === JaNei.NEI && (
-            <>
-              <Box maxWidth="60rem" paddingBlock="2 2">
-                <Textarea
-                  {...register('beskrivelse', {
-                    required: {
-                      value: true,
-                      message: 'Du må beskrive hvorfor bruker er unntatt fra aktivitetsplikten',
-                    },
-                  })}
-                  label="Beskrivelse"
-                  description="Beskriv hvorfor bruker er unntatt fra aktivitetsplikten"
-                  error={errors.beskrivelse?.message}
-                />
-              </Box>
-              <HStack gap="4">
-                <Button
-                  variant="secondary"
-                  size="small"
-                  disabled={isPending(lagreUnntakVarigStatus)}
-                  onClick={avbrytRedigering}
-                >
-                  Avbryt
-                </Button>
-                <Button loading={isPending(lagreUnntakVarigStatus)} variant="primary" type="submit" size="small">
-                  Lagre
-                </Button>
-              </HStack>
-            </>
-          )}
-        </VStack>
-      </form>
-
-      {jaHarAktivitetsplikt &&
-        (vurderingHarInnhold(vurdering) ? (
-          <HStack gap="4">
-            <Button variant="secondary" size="small" onClick={avbrytRedigering}>
-              Avbryt
-            </Button>
-            <Button size="small" variant="primary" onClick={avbrytRedigering}>
-              Lagre
-            </Button>
-          </HStack>
-        ) : (
-          <VurderAktivitetspliktWrapperBehandling
-            doedsdato={doedsdato}
-            behandling={behandling}
-            defaultOpen={true}
-            varigeUnntak={vurdering.unntak.filter(
-              (unntak) => unntak.unntak === AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT
-            )}
-          />
-        ))}
-    </>
-  ) : (
-    <VStack gap="4" maxWidth="32.5rem">
-      <div>
-        <Label size="medium">Har bruker aktivitetsplikt?</Label>
-        {jaHarAktivitetsplikt ? (
-          <BodyShort>Ja</BodyShort>
-        ) : (
-          <BodyShort>Nei, bruker er født i 1963 eller tidligere og har lav inntekt</BodyShort>
-        )}
-      </div>
-      {watch('harAktivitetsplikt') === JaNei.NEI && (
-        <div>
-          <Label>Begrunnelse</Label>
-          <BodyShort>{watch('beskrivelse')}</BodyShort>
-        </div>
-      )}
-      {redigerbar && (
-        <div>
-          <Button variant="secondary" size="small" onClick={() => setRedigerer(true)}>
-            Rediger
-          </Button>
-        </div>
-      )}
-    </VStack>
+export const finnVarigUnntak = (vurdering: IAktivitetspliktVurderingNyDto): IAktivitetspliktUnntak | undefined => {
+  return vurdering.unntak.find(
+    (unntak) => unntak.unntak === AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT
   )
 }
 
@@ -269,7 +83,7 @@ export const AktivitetspliktVurdering = ({
         </VStack>
 
         {vurdering &&
-          !harVarigUnntak(vurdering!!) &&
+          !finnVarigUnntak(vurdering) &&
           (vurdering.aktivitet.length > 0 || vurdering.unntak.length > 0) && (
             <>
               <AktivitetsgradOgUnntakTabellBehandling
@@ -283,9 +97,6 @@ export const AktivitetspliktVurdering = ({
                   doedsdato={doedsdato}
                   behandling={behandling}
                   defaultOpen={false}
-                  varigeUnntak={vurdering.unntak.filter(
-                    (unntak) => unntak.unntak === AktivitetspliktUnntakType.FOEDT_1963_ELLER_TIDLIGERE_OG_LAV_INNTEKT
-                  )}
                 />
               )}
             </>
